hackathon/requirements.txt
boto3>=1.28.0
requests>=2.31.0
pyyaml>=6.0
langchain>=0.0.267
openai>=1.3.0
anthropic>=0.5.0
strand>=0.1.0
mcp-client>=0.1.0
